+++ jest.config.ts
@@ -2,6 +2,11 @@ import type { Config } from 'jest';
 
 const config: Config = {
   preset: 'ts-jest',
+  globals: {
+    'ts-jest': {
+      tsconfig: 'tsconfig.jest.json',
+    },
+  },
   testEnvironment: 'node',
   testTimeout: 10000,
   roots: ['<rootDir>/tests', '<rootDir>/__tests__'],